Properly limit exception handling in requests
Rescuing from `Exception` is bad practice because is rescues EVERYTHING
in that block including `NoMemoryErrors` and `SyntaxErrors`. What this
means in practice is that cannot use gems like `webmock` and `fakeweb`
to test your integration. Limiting to `RestClient::Exception` will
rescue all HTTP-related errors, which is likely the intended behavior.

And alternative suggestion is to rescue `StandardError`, which is best
practice for "rescue everything" since broader issues do not inherit
from `StandardError`.