Remove console script, better to just use `bundle console`
This was confusing me for awhile and I was trying to get this to work
but realized it doesn’t make any sense. Bundler already takes care of
this for us. Since we are developing you should expect to run `bundle
install` and then just `bundle console` with an IRB console with the
gem loaded. Easy.